+++ merdog/include/word_record.hpp
@@ -54,7 +54,7 @@ namespace Mer
 	public:
 		StructRecorder() :WordRecorder(ESymbol::SSTRUCTURE)
 		{
-			type_code = Mem::type_counter+=2;
+			type_code = size_t(Mem::type_counter+=2);
 		}
 
 	};
